+++ Sources/BaseFoundationiOS/Common/GlobalTask.swift
@@ -11,7 +11,10 @@ public class GlobalTask {
     public static func performDelay(seconds: Double = 0.1, handler: @escaping SimpleHandler) {
         Task {
             try await Task.sleep(for: .seconds(seconds))
-            handler()
+            
+            await MainActor.run {
+                handler()
+            }
         }
     }
 }